requirements.txt
openai==0.27.0
aiogram==2.23.0
python-dotenv==1.0.0
asyncio==3.4.3
logging==0.5.1.2